+++ cmd/agent/main.go
@@ -2,14 +2,14 @@ package main
 
 import (
 	"fmt"
-	"github.com/kiraplenkin/go-musthave-devops/internal/sendingService"
+	"github.com/kiraplenkin/go-musthave-devops/internal/sender"
 	"github.com/kiraplenkin/go-musthave-devops/internal/types"
 	"time"
 )
 
 func main() {
-	sender := sendingService.NewSender()
-	logger := sendingService.NewLogger(*sender)
+	sender := sender.NewSender()
+	logger := sender.NewLogger(*sender)
 
 	for {
 		ticker := time.NewTicker(types.SenderConfig.ServerUpdateTime)
